local: Fix searching for MyDocs mmc device
Device name in /sys/block/ could not be same as in /dev/ if udev renaming it.
And this is real problem on Maemo without special kernel patch when device
booting with inserted SD card. So read major and minor numbers from /sys/
and find coresponding device in /dev/. These two numbers will be same after
renaming device too.